Refactor to make SoundSource stateless
This will allow things like repeating notes and effects like reversing
sounds easier to implement (I hope).

The Sequence class is given a repeat member that causes it to loop. This
requires either way to reset the state of component notes (difficult to
implement), or going stateless and pretending time has gone back to
zero.