Add caching to low pass filter
I had to make the SoundSource trait allow a &mut self to do this so all the
classes had to change.

Despite doing less calculations in realtime, the convolve function is
still too expensive (and I don't even know if I wrote it correctly).
I'll need to learn a bit more DSP to do this filter. I know there are
such things as crates but I'm using this as a learning exercise so I
want to at least write my own before converting it over to use a lib.