Fixed convolution
I'm pretty sure this is right now - there is no need to compute the
convolution for the full set of samples each time since whenever
next_value() is called, the samples array shifts by 1 sample and we just
do the inner convolution calcualtion for that sample.

Having said that, this low pass filter is not working very well so I
need to do more research. If seems to filter out almost everything
except for a small sliver of top end that i can barely hear.