Working low pass filter
I figured out the transition bandwidth has to be above the cutoff. I
previously thought that it added onto the cutoff but that is not the
case. I stil might have it wrong but it sounds like it's working the
way I have it now.

Also the filter gain seems better if I don't divide during the
convolution.