Refactor how patches are read from yaml
The rich_synthesis1.yaml is an example of the reworked format. I
realised that each patch needs to be self contained so it now has a
list of sound items embedded in the patch as it if were its own
isolated song but it will substitue parameters INPUT N with the param
from where the patch is called. It should also work nested because I
a stack to keep track of how many levels deep I am in the patches.