Add dc multiply param for convenience
Since it's cumbersome to set up a DC source just to make a multiply
param, allow it to be setup directly as an input. The yaml format is
already getting a bit crazy so i'm going to need to refactor but I'm
blazing a trail atm to get ideas.

I've also added a SoundSource that just takes an midi input and runs it
through the midi2freq function. That way we can use that func in the
yaml.