Add PATCH_INPUT source
Patches can now be refered with a source input that is different from a
parameter in that a SoundSource can be passed from the parent patch or
song into the patch as an input. This allows the patch to act like a
filter or to have a Knob style input.