Add wavetable
A wavetable is a basic sampler. Typically a ramp function is used to
lookup samples stored in the table and output them. The amplitude and
period of the ramp control the playback speed and how much of the sample
is played. A reverse ramp can be used to play the wavetable backwards or
a sinewave could be used to warp the sound.