Add dumb oscillator
The Oscillator class is different to Sine because it doesn't try and
control any phase issues. This is so I can play with frequency
modulation as per the example in the yaml file.

It might also support phase modulation more simply than other types that
use a ramp function to mess around with the phase. I'm not sure if that
input is useful yet.