Add cross fade formant example
Getting to understand the idea of formants now, this is a different
technique called cross fading. The misunderstnading I had before was
that I was expecting the FFT to show peak at the desired frequency but
it's more that if you draw an envelope over the FFT, the peak will be at
the formant frequency.