Add EXPR parser and PAF example
In order to realise phase aligned formant as a general patch, an
expression parser was added to make it possible do perform simple
functions of input parameters. These are done at the variable
substitution phase so are not performed per sample.

The EXPR() keyword is used to cause a subsitution to be made with the
evaluation of the expression inside the brackets. The INPUT
substitutions are made first so they can be used in the EXPR statements.

The INPUT syntax had to change to make it easier to parse them. They no
longer require a space before and after them but now need brackets so
to refer to param n, the syntax is INPUT(n).

NOTE: this breaks many of the yaml files which still use "INPUT n"
syntax. I'll fix those in a future commit.