Add recirculating delay
This is a major rework of the project to accomodate a new requirement.
The feedback required to implement recirculating delay meant the
introduction of BufferReader and BufferWriter. Before arriving at these
classes, I went to the effort to pull out all mutable data from the
SoundSources so that they are immutable again during playback. This is
to try and make it easier to reason about how the networks will behave.

I'm not sure I needed to do this for the reader/writer classes and may have
also violated my own principals in the implementation so consider this a
WIP patch.