Support buffer_reader and buffer_writer in yaml
There is a catch when using these. The buffer_writer must be reachable
from whatever root sound is chosen, otherwise only the reader is loaded
and the feedback loop will not be created.

See the example recirculating_delay_patch.yaml and the example that uses
it recirculating_delay_impluse1.yaml. If 'mix' is chosen as the patch
root, then write_buf is not referenced so it will never be created.
Making write_buf the root ensures it is referenced.