Add ImportWav
This allows external sounds in the form of wav files to be imported into
the songs. There is a convenience command line arg so you can play a wav
file as well.

I've used the wavetable approach to playback so that the input sample
rate doesn't have to match the output sample rate and interpolation will
be used. A future version of this could allow import to a WaveTable
specifying the sweep function separately so the playback can be
stretched and cropped etc...