Change PitchShift to use PreRender
The original implementation using BufferWriter and  BufferReader had
some issues where I was unable to determine exactly how it was working
at runtime since the buffer is written while playing and is intended for
feedback style effects rather than reusing a buffer.

To get around this, I instead prerender the input and clone it.