Add PulseTrain
We are about to add some filter functions and I'll want to get transfer
functions so a pulse train generator should come in handy.

Hot Tip: Audacity is a free audio editor that can view wav files and
has analyse view which shows the FFT of a selected part of your file so
running a pulse train through your filter and then loading the result
in audacity saves you buying matlab or whatever.