Add ElementaryNonRecirculatingFilter
I'm not sure if I got this right but it's my best attempt at the most
basic of filters. It's not very useful but is a building block for
more complex filters. By using a single sample delay, it causes the gain
to roll off from 1 at 0Hz to 0 or -infinity dB at the nyquist frequency.

I have my doubts about whether I understood the complex gain properly. I
suspect it will be more important once there is feedback involved.