Add dyn-clone
This allows the SoundSource trait to be cloneable and also object safe
meaning that the dynamic binding stuff can still work. Usually a trait
is used with dyn can not be Sized and the standard Clone trait add Sized
to the trait. To get around this, the dyn-clone trait does some arcane
rust magic.