Add complex sound source for filters
When chaining elementary filters together, the complex signal is needed
for processing by each filter in the chain rather than converting to a
real signal at each stage. A new PoleZeroFilter has been introduced and
a RealToComplex sound source so that I don't have to make every signal
in the whole project convert to complex numbers.