Add ButterworthBandpassFilter
I don't think this is quite right and I'm not sure what this filter is
for but this is the best I could do. It puts a very narrow spike at the
centre frequency. The original bandwidth of the low pass filter seems
to have very little influence on the shape of the filter.