Add simpler LowPassFilter
This is closer to the preset lop~ filter in puredata. It is a single
pole filter that takes two signal inputs: the input signal and the
cutoff freq signal. This makes it useful in filter sweeps and other
dynamic effects.

The previous filter that was based on a more post-processing approach
has been removed.

I also had to change the TimeBox effect for use in this class so that
the cutoff signal is the same length as the input signal.