Fix issue with nested patches and PATCH_INPUT
When reading a patch that used a PATCH_INPUT which itself uses a patch,
the context object was being manipulated in a non-reentrant way so that
the second attempt to get a patch would fail. This solution avoids
temporarily altering the context so that patches within PATCH_INPUTs
will be handled.